Issue #SB-14104 fix: removed indexing creation logic and replaced with index template
