+++ platform-modules/batch-models/src/main/scala/org/ekstep/analytics/job/CourseMetricsJob.scala
@@ -16,7 +16,6 @@ import org.joda.time.format.DateTimeFormat
 import org.sunbird.cloud.storage.conf.AppConf
 import org.sunbird.cloud.storage.factory.{StorageConfig, StorageServiceFactory}
 
-import scala.collection.mutable.ListBuffer
 import scala.collection.{Map, _}
 
 trait ReportGenerator {
@@ -29,7 +28,7 @@ trait ReportGenerator {
   def saveReport(reportDF: DataFrame, url: String): Unit
 }
 
-case class ESIndexResponse(isIndexCreated: Boolean, isOldIndexDeleted: Boolean, isIndexLinkedToAlias: Boolean)
+case class ESIndexResponse(isOldIndexDeleted: Boolean, isIndexLinkedToAlias: Boolean)
 
 object CourseMetricsJob extends optional.Application with IJob with ReportGenerator {
 
@@ -270,29 +269,25 @@ object CourseMetricsJob extends optional.Application with IJob with ReportGenera
 
     val cBatchIndex = AppConf.getConfig("course.metrics.es.index.cbatch")
     val aliasName = AppConf.getConfig("course.metrics.es.alias")
-    val cBatchStatsIndex = getFormatedIndexName()
+    val newIndexPrefix = AppConf.getConfig("course.metrics.es.index.cbatchstats.prefix")
+    val newIndex = suffixDate(newIndexPrefix)
     try {
-      val response: ESIndexResponse = createEsIndex(cBatchStatsIndex, aliasName)
-      val indexList = getIndexName(aliasName)
-      val index = indexList.mkString("")
-      if(index.nonEmpty){
-        batchStatsDF.saveToEs(s"$index/_doc", Map("es.mapping.id" -> "id"))
-        JobLogger.log("Indexing batchStatsDF is success: " + index, None, INFO)
-      }else{
-        JobLogger.log("Indexing batchStatsDF is got failed: " + index, None, ERROR)
-      }
+      val indexList = ESUtil.getIndexName(aliasName)
+      val oldIndex = indexList.mkString("")
+      batchStatsDF.saveToEs(s"$newIndex/_doc", Map("es.mapping.id" -> "id"))
+      JobLogger.log("Indexing batchStatsDF is success: " + newIndex, None, INFO)
 
-
+      if (!oldIndex.equals(newIndex)) ESUtil.rolloverIndex(newIndex, aliasName)
 
       // upsert batch details to cbatch index
-     batchDetailsDF.saveToEs(s"$cBatchIndex/_doc", Map("es.mapping.id" -> "id"))
-    val batchStatsPerBatchCount = batchStatsDF.groupBy("batchId").count().collect().map(_.toSeq)
-    val batchStatsCount  = batchStatsDF.count()
+      batchDetailsDF.saveToEs(s"$cBatchIndex/_doc", Map("es.mapping.id" -> "id"))
+      val batchStatsPerBatchCount = batchStatsDF.groupBy("batchId").count().collect().map(_.toSeq)
+      val batchStatsCount  = batchStatsDF.count()
 
-    val batchDetailsPerBatchCount = batchDetailsDF.groupBy("id").count().collect().map(_.toSeq)
-    val batchDetailsCount = batchDetailsDF.count()
+      val batchDetailsPerBatchCount = batchDetailsDF.groupBy("id").count().collect().map(_.toSeq)
+      val batchDetailsCount = batchDetailsDF.count()
 
-    JobLogger.log(s"CourseMetricsJob: Elasticsearch index stats { $cBatchStatsIndex : { perBatchCount: ${JSONUtils.serialize(batchStatsPerBatchCount)}, totalNoOfRecords: $batchStatsCount }, $cBatchIndex: { perBatchCount: ${JSONUtils.serialize(batchDetailsPerBatchCount)}, totalNoOfRecords: $batchDetailsCount } }", None, INFO)
+      JobLogger.log(s"CourseMetricsJob: Elasticsearch index stats { $newIndex : { perBatchCount: ${JSONUtils.serialize(batchStatsPerBatchCount)}, totalNoOfRecords: $batchStatsCount }, $cBatchIndex: { perBatchCount: ${JSONUtils.serialize(batchDetailsPerBatchCount)}, totalNoOfRecords: $batchDetailsCount } }", None, INFO)
 
     } catch {
       case ex: Exception => {
@@ -302,48 +297,8 @@ object CourseMetricsJob extends optional.Application with IJob with ReportGenera
     }
   }
 
-  def getFormatedIndexName(): String = {
-    val date: String = DateTimeFormat.forPattern("dd-MM-yyyy").print(DateTime.now())
-    val indexPrefixName = AppConf.getConfig("course.metrics.es.index.cbatchstats.prefix")
-    indexPrefixName + date
-  }
-
-  def createEsIndex(indexName: String, aliasName: String): ESIndexResponse = {
-    val createIndexResponse = ESUtil.createIndex(indexName, AppConf.getConfig("course.metrics.es.mapping"))
-    val olderIndexList = getIndexName(aliasName)
-    if (createIndexResponse.acknowledged) {
-      JobLogger.log("Newly created  index name is:" + indexName + " and " + "Alias Name is:" + aliasName, Option(indexName), INFO)
-      val addIndexToAliasResponse = ESUtil.addIndexToAlias(indexName, aliasName)
-      if (addIndexToAliasResponse.acknowledged && olderIndexList.nonEmpty) {
-        JobLogger.log("Adding index (" + indexName + ") to alias(" + aliasName + ") is success", None, INFO)
-        val deleteIndexResponse = ESUtil.deleteIndex(olderIndexList)
-        if (deleteIndexResponse.acknowledged) {
-          JobLogger.log("Delete index is success! Index Name: " + olderIndexList, None, INFO)
-          ESIndexResponse(createIndexResponse.acknowledged, deleteIndexResponse.acknowledged, addIndexToAliasResponse.acknowledged)
-        } else {
-          JobLogger.log("Delete index is failed! Index Name: " + olderIndexList, None, ERROR)
-          JobLogger.log(deleteIndexResponse.toString, None, ERROR)
-          ESIndexResponse(createIndexResponse.acknowledged, deleteIndexResponse.acknowledged, addIndexToAliasResponse.acknowledged)
-        }
-      } else {
-        JobLogger.log("Adding alias " + aliasName + " to index " + indexName + " status is: " + addIndexToAliasResponse.acknowledged, None, ERROR)
-        JobLogger.log(addIndexToAliasResponse.toString, None, ERROR)
-        ESIndexResponse(createIndexResponse.acknowledged, false, addIndexToAliasResponse.acknowledged)
-      }
-    } else {
-      JobLogger.log("Create Elastic search index " + indexName + " is got failed", None, ERROR)
-      JobLogger.log(createIndexResponse.toString, None, ERROR)
-      ESIndexResponse(createIndexResponse.acknowledged, false, false)
-    }
-  }
-
-  def getIndexName(aliasName: String): List[String] = {
-    val indexMap = ESUtil.listIndexByAlias(aliasName)
-    var indexListBuffer = new ListBuffer[String]()
-    indexMap.foreach(element => {
-      indexListBuffer += element.getOrElse("index", null)
-    })
-    indexListBuffer.toList
+  def suffixDate(index: String): String = {
+    index + DateTimeFormat.forPattern("dd-MM-yyyy").print(DateTime.now())
   }
 
   def saveReport(reportDF: DataFrame, url: String): Unit = {
